--- towhee/models/layers/__init__.py
@@ -1,8 +1,0 @@
-from .weight_init import trunc_normal_, variance_scaling_, lecun_normal_
-from .mlp import Mlp
-from .convmlp import ConvMlp
-from .gatedmlp import GatedMlp
-from .dropblock2d import DropBlock2d, drop_block_2d
-from .droppath import DropPath
-
-
