+++ towhee/models/swin_transformer/utils.py
@@ -11,46 +11,40 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License
+
+
 from torch import nn
-from itertools import repeat
-import collections.abc
-from towhee.trainer.models.layers import trunc_normal_, lecun_normal_
+from towhee.models.utils.weight_init import trunc_normal_, lecun_normal_
+
 
 def window_partition(x, window_size: int):
     """
     Args:
-        x: (b, h, w, c)
-        window_size (int): window size
+       x: (b, h, w, c)
+       window_size (int): window size
     Returns:
-        windows: (num_windows*B, window_size, window_size, c)
+       windows: (num_windows*B, window_size, window_size, c)
     """
     b, h, w, c = x.shape
     x = x.view(b, h // window_size, window_size, w // window_size, window_size, c)
     windows = x.permute(0, 1, 3, 2, 4, 5).contiguous().view(-1, window_size, window_size, c)
     return windows
 
+
 def window_reverse(windows, window_size: int, h: int, w: int):
     """
     Args:
-        windows: (num_windows*b, window_size, window_size, c)
-        window_size (int): Window size
-        h (int): Height of image
-        w (int): Width of image
+       windows: (num_windows*b, window_size, window_size, c)
+       window_size (int): Window size
+       h (int): Height of image
+       w (int): Width of image
     Returns:
-        x: (b, h, w, c)
+       x: (b, h, w, c)
     """
     b = int(windows.shape[0] / (h * w / window_size / window_size))
     x = windows.view(b, h // window_size, w // window_size, window_size, window_size, -1)
     x = x.permute(0, 1, 3, 2, 4, 5).contiguous().view(b, h, w, -1)
     return x
-
-# From PyTorch internals
-def _ntuple(n):
-    def parse(x):
-        if isinstance(x, collections.abc.Iterable):
-            return x
-        return tuple(repeat(x, n))
-    return parse
 
 def init_vit_weights(module: nn.Module, name: str = '', head_bias: float = 0., jax_impl: bool = False):
     """ ViT weight initialization
@@ -85,9 +79,3 @@ def init_vit_weights(module: nn.Module, name: str = '', head_bias: float = 0., j
     elif isinstance(module, (nn.LayerNorm, nn.GroupNorm, nn.BatchNorm2d)):
         nn.init.zeros_(module.bias)
         nn.init.ones_(module.weight)
-
-to_1tuple = _ntuple(1)
-to_2tuple = _ntuple(2)
-to_3tuple = _ntuple(3)
-to_4tuple = _ntuple(4)
-to_ntuple = _ntuple
