+++ towhee/models/utils/weight_init.py
