--- towhee/trainer/models/layers/__init__.py
@@ -1,4 +1,0 @@
-from .weight_init import trunc_normal_, variance_scaling_, lecun_normal_
-from .mlp import Mlp, GluMlp, GatedMlp
-from .drop import DropBlock2d, DropPath, drop_block_2d
-
